+++ src/BoxStruct.cpp
@@ -2,7 +2,7 @@
 
 Box::Box() : Box(0, 0, 0, 0, 0) {}
 
-Box::Box(const int newLength, const int newWidth, const int newHeight, const double newWeight, const int newValue)
+Box::Box(const int newLength, const int newWidth, const int newHeight, const double newWeight, const double newValue)
         : length(newLength),
           width(newWidth),
           height(newHeight),
